phase_3_visualization/requirements.txt
streamlit==1.28.0
plotly==5.15.0
pandas==2.0.3
cassandra-driver==3.28.0
